feat: contract factory deploy arguments
